Adding depth to the rechit

+++ DataFormats/ParticleFlowReco/interface/PFRecHit.h
@@ -73,7 +73,7 @@ namespace reco {
     void addNeighbour(short x,short y, short z,const PFRecHitRef&);
     const PFRecHitRef getNeighbour(short x,short y, short z);
     void setTime( double time) { time_ = time; }
-    void setDepth( unsigned short depth) { depth_ = depth; }
+    void setDepth( int depth) { depth_ = depth; }
     void clearNeighbours() {
       neighbours_.clear();
     }
@@ -113,7 +113,7 @@ namespace reco {
     double time() const { return time_; }
 
     /// depth for segemntation
-    unsigned short  depth() const { return depth_; }
+    int  depth() const { return depth_; }
 
     /// rechit momentum transverse to the beam, squared.
     double pt2() const { return energy_ * energy_ *
@@ -184,7 +184,7 @@ namespace reco {
 
 
     /// depth
-    unsigned short      depth_;
+    int      depth_;
 
     /// rechit cell centre: x, y, z
     math::XYZPoint      position_;
